Use {name}, rather than {feed}, for alt text.
As a screen reader user, the previous behavior is spammy because it reads the key contents to me rather than the poster's profile name. This makes the experience of reading my feed far more pleasant. It looks like {feed} is the fallback, so I don't think this should break anything.