[[ calendar ]] initial support for exporting gatherings to ICS
It is working and the iCalendar files generated passed all validation tests but I'm facing some trouble when trying to import them in either Thunderbird Calendar or Windows 10 Calendar, regarding the _attendees_ fields. They don't appear on windows 10 and appear with extra attributes on Thunderbird. Quite odd, but the event still works, so I'm shipping.