[[ packages ]] changes to how packages are loaded
This changes how cards are selected for rendering cards in a collection of messages.

Now, multiple packages can declare they can handle the same message type.

They need to pass a `validator` function in the message type declaration.

All validators are run for a given message. The first one to return true becomes the card.