[[ calendar ]] add missing about handlers
Adds handlers for dealing with `about` messages related to gatherings.

It adds _changed the image for_ in case the user sets the image for a gathering and a catch all case _made changes to_ in the case we have no idea what is going on but we know it is gathering related.

closes: #61