[[ webextension ]] adding browser polyfill.
This fixes the lack of `browser.*` API namespace on chromium-based browsers. Patchfox had the polyfill vendored and loaded in some pages, but that was not enough to add it to all the places it needed to be added.

I've switched from the vendored copy to a NPM module for the polyfill. Added a task to copy it around, and added it  to many places where it was needed.