[[ settings ]] removes `columnSize` preference.
The `columnSize` preference was a left-over from the first move to use `augmented-ui`. It allowed the user to toggle between the narrow column used by the `augmented-ui` layout and a larger layout that filled the whole page.

For a while, the short column has been forced in `wm.svelte` regardless of the preference setting. This made the toggle in _Settings &rarr; Display Preferences_ useless.

That toggle has been removed.

closes: #93