[[ ssb ]] refactored `ssb.sbot.id` to `ssb.feed`.
This is part of #104: identifying all direct usage of sbot.

Some of that usage is easy to refactor right away, such as cases where `sbot.id` is used to get the feed id for the running user. Patchfox has always exposed the feed id as `ssb.feed`, it was just a matter of using it.